requirements.txt
bcrypt==4.1.1
flask==2.3.3
SQLAlchemy==2.0.20
PyJWT==2.3.0
pydantic==1.9.2